新增时间轮定时器

+++ utilities/hashed-wheel-timer/src/test/java/com/silong/foundation/utilities/hwtimer/HashedWheelTimerTests.java
@@ -39,7 +39,7 @@ public class HashedWheelTimerTests {
     HashedWheelTimer timer = new HashedWheelTimer();
     int result = 1;
     try (DelayedTask a = timer.submit("A", () -> result, 0, TimeUnit.HOURS)) {
-      int num = (int) a.getResult();
+      int num = a.getResult();
       Assertions.assertEquals(result, num);
     }
     Assertions.assertTrue(timer.stop());
